restructured the footer
