+++ src/assets/css/styles.css
@@ -1,3 +1,51 @@
+/* -------footer css-------------------------- */
+li {
+    list-style-type: none;
+    padding-right: 4%;
+}
+footer {
+    background-color: #2c2c2c;
+    position: absolute;
+    bottom: -0;
+    width: 100%;
+    overflow: hidden;
+    color: white
+}
+img {
+    width: 6%;
+    float: left;
+    padding-right: 3%;
+    padding-bottom: 3%;
+}
+@media only screen and(min-width:600px) {
+    li {
+        list-style-type: none;
+        float: left;
+        padding-right: 6%;
+        margin-top: 1.5%;
+        margin-left: 3%;
+    }
+    img {
+        width: 2rem;
+        float: left;
+        padding-bottom: 30%;
+        position: absolute;
+
+    }
+
+    footer {
+        background-color: #2c2c2c;
+        position: absolute;
+        bottom: -0;
+        width: 100%;
+        height: 13%;
+        overflow: hidden;
+        color: white
+    }
+}
+
+
+
 /* -----------App header css---------- */
 .navbar-custom {
     height: 15vh
@@ -47,6 +95,10 @@
 .dropdown-section> .dropdown> .dropdown-toggle::after{
     margin-left: 4em;
 }
+.navbar-custom> .navbar-brand> img{
+    width: auto;
+}
+
 
 /* ------------custom css of black nevigation bar---------- */
 .dark-nav-bar> .navbar {
