+++ src/components/index.js
@@ -3,8 +3,7 @@ import { Router, Route, Switch } from "react-router-dom";
 import { history } from "../store";
 import AppHeader from "./static/NavBar";
 import AppNavBar from "./static/NavBar/appNavBar";
-import AppHeader from "./NavBar";
-import Footer from "./footerComponent/footer";
+import Footer from "./static/Footer";
 const HomePage = lazy(() => import("./views/homepage"));
 const Facilities = lazy(() => import("./views/facilities"));
