+++ src/index.js
@@ -1,9 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./assets/css/styles.css";
-// import "./assets/css/material-dashboard-react.css?v=1.8.0";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
